Add 2021 Day 11 files
Also had to re-org the ContentView a little because a Group can only hold 10 subviews and none of those subviews can be another Group. So they're broken up into groups of 5 days now and that should cover it. I also made the current year be organized with the most recent entry at the top... some day this should be rebuilt to not need to be a manual process... some day.