Pull in swift-collections and use a Heap to make it way faster
This required some refactoring, but definitely fixed the performance problems. I'm printing out the number of Points processed too, because without the estimatedCostToEnd was ever so slightly faster and it turns out on the input it processes _more_ points. So I'm gonna see how that goes on Part 2, but it seems like Djikstra's might actually be faster than A* for this, which is surprising.