Only run the update-readme step on PR merged (and on workflow-dispatch)
I saw an update to the template repo that mentioned everyone running the cron at the same time was breaking it so they requested people change the minute on the schedule. I don't really want it scheduled, I'd rather it ran when I update anyway. And I like having PRs in the history for each day anyway.