Solve day 11 part 2
I totally had to find help with the math trick :P

Multiply all the divisible_by's together and then mod the worry levels with that before doing any math on them. I called it LCM for Lowest Common Multiple... which is not what it _is_, it's what it helps you find, I think?