Actually solve day 17 part 2
Took a break and went and looked at r/AdventOfCode for hints and found other people with similar problems to mine and there seemed to be a common thread of increasing the lookback for the pattern detection for the real input. I refactored that bit of code to make that a bit more configurable and it worked!