Solve day 19 part 2
In order to make this work in a reasonable amount of time, had to find a few more optimization points. Previously it was checking if the number of bots vs the max spend amount for that type of bot, now it checks number of bots plus amount of that resource in storage vs the amount that could be spent in the remaining time.

That's a pretty big performance boost, but I think the bigger one is passing in a time_map which keeps track of the best result at every minute of the duration. If an iteration doesn't have as many as the best one for its current minute, it bails early because it'll never catch up.